# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Felt API Doc",
        "description": "Felt API Documentation",
        "contact": {
            "email": "rejohnbd@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customer/profile": {
            "get": {
                "tags": [
                    "getCustomerProfile"
                ],
                "summary": "Returns Customer Profile",
                "description": "Returns Customer Profile Information",
                "operationId": "getCustomerProfile",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "updateCustomerProfile"
                ],
                "summary": "Updated Customer",
                "description": "Update Customer User Profile.",
                "operationId": "updateCustomerProfile",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "path",
                        "description": "First Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "path",
                        "description": "Last Name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_optional",
                        "in": "path",
                        "description": "Optional Phone",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "email_optional",
                        "in": "path",
                        "description": "Optional Email",
                        "schema": {
                            "type": "email"
                        }
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Address",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/customer/update-password": {
            "post": {
                "tags": [
                    "updateCustomerPassword"
                ],
                "summary": "Updated Customer",
                "description": "Update Customer User Profile.",
                "operationId": "updateCustomerPassword",
                "parameters": [
                    {
                        "name": "old_password",
                        "in": "path",
                        "description": "Old Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "path",
                        "description": "New Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "confirm_new_password",
                        "in": "path",
                        "description": "Confirm New Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/vehicle-papers": {
            "get": {
                "tags": [
                    "vehicle-papers"
                ],
                "summary": "Returns all vehicle-papers",
                "description": "All vehicle papers. This is GET api common for both user Admin & Customer user",
                "operationId": "vehicle-papers",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}